fixed valid dtype for tests

+++ ivy_tests/test_ivy/test_frontends/test_tensorflow/test_tensor.py
@@ -82,7 +82,7 @@ def test_tensorflow_instance_add(dtype_and_x, as_variable, native_array):
 @handle_cmd_line_args
 @given(
     dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
+        available_dtypes=helpers.get_dtypes("float"),
         num_arrays=2,
         shared_dtype=True,
     ),
@@ -465,7 +465,7 @@ def test_tensorflow_instance_radd(dtype_and_x, as_variable, native_array):
 @handle_cmd_line_args
 @given(
     dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
+        available_dtypes=helpers.get_dtypes("float"),
         num_arrays=2,
         shared_dtype=True,
     ),
@@ -529,7 +529,7 @@ def test_tensorflow_instance_rfloordiv(dtype_and_x, as_variable, native_array):
 @handle_cmd_line_args
 @given(
     dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
+        available_dtypes=helpers.get_dtypes("float"),
         num_arrays=2,
         shared_dtype=True,
     ),
